+++ includes/Maps_MappingService.php
@@ -45,7 +45,7 @@ abstract class MapsMappingService implements iMappingService {
 	 *
 	 * @var array
 	 */
-	protected $resourceModules = array();
+	protected $resourceModules = [];
 
 	/**
 	 * A list of dependencies (header items) that have been added.
@@ -54,7 +54,7 @@ abstract class MapsMappingService implements iMappingService {
 	 *
 	 * @var array
 	 */
-	private $addedDependencies = array();
+	private $addedDependencies = [];
 
 	/**
 	 * A list of dependencies (header items) that need to be added.
@@ -63,7 +63,7 @@ abstract class MapsMappingService implements iMappingService {
 	 *
 	 * @var array
 	 */
-	private $dependencies = array();
+	private $dependencies = [];
 
 	/**
 	 * Constructor. Creates a new instance of MapsMappingService.
@@ -73,7 +73,7 @@ abstract class MapsMappingService implements iMappingService {
 	 * @param string $serviceName
 	 * @param array $aliases
 	 */
-	public function __construct( $serviceName, array $aliases = array() ) {
+	public function __construct( $serviceName, array $aliases = [] ) {
 		$this->serviceName = $serviceName;
 		$this->aliases = $aliases;
 	}
@@ -130,7 +130,7 @@ abstract class MapsMappingService implements iMappingService {
 	 * @return array
 	 */
 	public function getConfigVariables() {
-		return array();
+		return [];
 	}
 
 	/**
@@ -140,7 +140,7 @@ abstract class MapsMappingService implements iMappingService {
 	 */
 	public final function getDependencyHtml() {
 		$allDependencies = array_merge( $this->getDependencies(), $this->dependencies );
-		$dependencies = array();
+		$dependencies = [];
 
 		// Only add dependnecies that have not yet been added.
 		foreach ( $allDependencies as $dependency ) {
@@ -151,7 +151,7 @@ abstract class MapsMappingService implements iMappingService {
 		}
 
 		// If there are dependencies, put them all together in a string, otherwise return false.
-		return $dependencies !== array() ? implode( '', $dependencies ) : false;
+		return $dependencies !== [] ? implode( '', $dependencies ) : false;
 	}
 
 	/**
@@ -162,7 +162,7 @@ abstract class MapsMappingService implements iMappingService {
 	 * @return array
 	 */
 	protected function getDependencies() {
-		return array();
+		return [];
 	}
 
 	/**