+++ includes/geocoders/Maps_GoogleGeocoder.php
@@ -43,7 +43,7 @@ final class MapsGoogleGeocoder extends \Maps\Geocoder {
 	 * 
 	 * @since 0.7
 	 * 
-	 * @param string $address
+	 * @param string $response
 	 * 
 	 * @return array
 	 */		
@@ -54,10 +54,10 @@ final class MapsGoogleGeocoder extends \Maps\Geocoder {
 		// In case on of the values is not found, return false.
 		if ( !$lon || !$lat ) return false;
 
-		return array(
+		return [
 			'lat' => (float)$lat,
 			'lon' => (float)$lon
-		);
+		];
 	}
 	
 	/**
@@ -68,7 +68,7 @@ final class MapsGoogleGeocoder extends \Maps\Geocoder {
 	 * @return array
 	 */
 	public static function getOverrides() {
-		return array( 'googlemaps3' );
+		return [ 'googlemaps3' ];
 	}
 	
 }