+++ src/Maps/Elements/WmsOverlay.php
@@ -23,7 +23,7 @@ class WmsOverlay extends BaseElement {
 	private $wmsLayerName;
 
 	/**
-	 * @var String WMS Stype name (default value: 'default')
+	 * @var String WMS Style name (default value: 'default')
 	 */
 	private $wmsStyleName;
 
@@ -96,11 +96,11 @@ class WmsOverlay extends BaseElement {
 	public function getJSONObject ( $defText = "", $defTitle = "" ) {
 		$parentArray = parent::getJSONObject( $defText , $defTitle );
 
-		$array = array (
+		$array =  [
 			'wmsServerUrl' => $this->getWmsServerUrl() ,
 			'wmsLayerName' => $this->getWmsLayerName() ,
 			'wmsStyleName' => $this->getWmsStyleName()
-		);
+		];
 		return array_merge( $parentArray, $array );
 	}