+++ tests/phpunit/parserhooks/CoordinatesTest.php
@@ -32,51 +32,51 @@ class CoordinatesTest extends ParserHookTest {
 	 * @return array
 	 */
 	public function parametersProvider() {
-		$paramLists = array();
+		$paramLists = [];
 
-		$paramLists[] = array(
-			array(
+		$paramLists[] = [
+			[
 				'location' => '4,2',
 				'format' => 'dms',
 				'directional' => 'no',
-			),
+			],
 			'4° 0\' 0.00", 2° 0\' 0.00"'
-		);
+		];
 
-		$paramLists[] = array(
-			array(
+		$paramLists[] = [
+			[
 				'location' => '55 S, 37.6176330 W',
 				'format' => 'dms',
 				'directional' => 'no',
-			),
+			],
 			'-55° 0\' 0.00", -37° 37\' 3.48"'
-		);
+		];
 
-		$paramLists[] = array(
-			array(
+		$paramLists[] = [
+			[
 				'location' => '4,2',
 				'format' => 'float',
 				'directional' => 'no',
-			),
+			],
 			'4, 2'
-		);
+		];
 
-		$paramLists[] = array(
-			array(
+		$paramLists[] = [
+			[
 				'location' => '-4,-2',
 				'format' => 'float',
 				'directional' => 'yes',
-			),
+			],
 			'4 S, 2 W'
-		);
+		];
 
-		$paramLists[] = array(
-			array(
+		$paramLists[] = [
+			[
 				'location' => '55 S, 37.6176330 W',
 				'directional' => 'yes',
-			),
+			],
 			'55° 0\' 0.00" S, 37° 37\' 3.48" W'
-		);
+		];
 
 		return $paramLists;
 	}
@@ -88,45 +88,45 @@ class CoordinatesTest extends ParserHookTest {
 	 */
 	public function processingProvider() {
 		$definitions = ParamDefinition::getCleanDefinitions( $this->getInstance()->getParamDefinitions() );
-		$argLists = array();
+		$argLists = [];
 
-		$values = array(
+		$values = [
 			'location' => '4,2',
-		);
+		];
 
-		$expected = array(
+		$expected = [
 			'location' => new LatLongValue( 4, 2 ),
-		);
+		];
 
-		$argLists[] = array( $values, $expected );
+		$argLists[] = [ $values, $expected ];
 
-		$values = array(
+		$values = [
 			'location' => '4,2',
 			'directional' => $definitions['directional']->getDefault() ? 'no' : 'yes',
 			'format' => 'dd',
-		);
+		];
 
-		$expected = array(
+		$expected = [
 			'location' => new LatLongValue( 4, 2 ),
 			'directional' => !$definitions['directional']->getDefault(),
 			'format' => Maps_COORDS_DD,
-		);
+		];
 
-		$argLists[] = array( $values, $expected );
+		$argLists[] = [ $values, $expected ];
 
-		$values = array(
+		$values = [
 			'location' => '4,2',
 			'directional' => $definitions['directional']->getDefault() ? 'NO' : 'YES',
 			'format' => ' DD ',
-		);
+		];
 
-		$expected = array(
+		$expected = [
 			'location' => new LatLongValue( 4, 2 ),
 			'directional' => !$definitions['directional']->getDefault(),
 			'format' => Maps_COORDS_DD,
-		);
+		];
 
-		$argLists[] = array( $values, $expected );
+		$argLists[] = [ $values, $expected ];
 
 		return $argLists;
 	}